Gamma value adjusted as my own liking
Reverted back to google default for colors at 255 each.

And changed gamma as per my own liking.
Gamma is in hex values with their decimal converted range of 0 to 31

In amp0
Red changed 0x00 to 0x04 (0 to 4)
Green as default
Blue changed 0x1F to 0x1A (31 to 26)

In amp1
Red changed from 0x00 to 0x02 (0 to 2)
Green as default
Blue as default

This is dangerous so please try to keep it below 31 (decimal) and everything below 0x1F in hex

Signed-off-by: Harsh Bhanvadia <harsh.slade@gmail.com>